Move version badge to mod version
